Update build to use Gradle 1.11 and remain compatible with Guava 16
Gradle 1.9 and later appears to be rather sensitive when building
code that depends upon Guava, particularly if there are multiple
versions of Guava involved. This commit configures a resolution
strategy to ensure that all configurations use the specified version
of Guava, rather than using whatever version is pulled in
transitively. This avoids the problem described in GRADLE-2962 [1].

There still appears to be an underlying problem that may affect us in
the future. Even with the above-described resolution strategy in
place, running a build with -verbose:class shows that Guava classes
are being loaded from the Guava jar that Gradle itself provides
(guava-jdk5-14.0.1), not the one that's specified by the dependency
declared in the build. I have an issue open with Gradleware to address
this.

[1] http://issues.gradle.org/browse/GRADLE-2962