Add the Spring IO plugin
Configure the Spring IO plugin such that it's only applied when the
build is run with -PplatformVersion=<version>. This platformVersion
property is used to determine the version of the Platform that will
be used when running the springIoCheck task. The plugin can be used
by running a build as follows:

./gradlew clean build springIoCheck -PplatformVersion=1.0.0.BUILD-SNAPSHOT -PJDK7_HOME=…

This will test the project on JDK7 using the dependencies defined in
the latest snapshot of Spring IO Platform 1.0.0.

Gradle has been downgraded to 1.8 to avoid problems with an
incompatibility between Guava and Google Collections.